Move planar_contour out of extension
